+++ Project/Character.h
@@ -8,7 +8,7 @@
 class Character : public Actor
 {
 public:
-	Character(GameStruct::Box box, float health, float speed, bool isPossessed, GameStruct::vector2 forwardVector, int ID);
+	Character(GameStruct::Box box, int health, float speed, bool isPossessed, GameStruct::vector2 forwardVector, int ID);
 	virtual ~Character();
 
 	// -------------------------
